Added Y sorting and fixed collision bugs in the Mover component.

+++ ECS/Components/Graphics/BoxRenderer.cs
@@ -18,7 +18,7 @@ namespace FedoraEngine.ECS.Components.Graphics
             get => new Rectangle((int)Position.X, (int)Position.Y, Bounds.Width * (int)Scale.X, Bounds.Height * (int)Scale.Y);
         }
 
-        public int RenderLayer { get; set; }
+        public float RenderLayer { get; set; }
 
         public BoxRenderer(Rectangle rect, Color colour)
         {
